news_api/requirements.txt
fastapi
uvicorn
requests
pymongo
dnspython
pytest
pytest-mongo
pytest-cov
pytest-sugar
pyjwt
passlib
bcrypt
python-multipart